Вынесли Dekanat отдельно
